connectors/steps/bigquery/requirements.txt
google-cloud-bigquery==3.7.0
valohai-utils==0.3.1
tabulate==0.9.0
